Move: hexapod/plotter to templates/plotter
Move: hexapod/plotter to templates/plotter

Move: hexapod/plotter to templates/plotter

HexapodPlot fix
