+++ src/hexapod/index.js
@@ -1,5 +1,5 @@
 import VirtualHexapod from "./VirtualHexapod"
 import solveInverseKinematics from "./solvers/ik/hexapodSolver"
-import getNewPlotParams from "./plotter"
+import getNewPlotParams from "../templates/plotter"
 import { POSITION_NAMES_LIST } from "./constants"
 export { VirtualHexapod, getNewPlotParams, solveInverseKinematics, POSITION_NAMES_LIST }
